requirements.txt
django
djangorestframework
Pillow
drf-oidc-auth
gunicorn
requests